requirements.txt
Django==5.1
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2==2.9.5
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1